requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
openai==0.28
pinecone-client
python-dotenv